requirements.txt
aiohttp==3.7.4.post0
amqp==1.4.9
anyjson==0.3.3
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
billiard==3.3.0.23
celery==3.1.18
chardet==4.0.0
Django==3.2.6
djangorestframework==3.12.4
idna==3.2
kombu==3.0.37
multidict==5.1.0
Pillow==8.3.1
pytz==2021.1
reportlab==3.6.1
sqlparse==0.4.1
typing-extensions==3.10.0.2
yarl==1.6.3
